Sweep the last five bugprone checks and make the tier binding
The five checks the pre-push gate still excluded are at zero, so
WarningsAsErrors carries 'bugprone-*,cert-*' from here.

Seven of the 19 widening findings were literal products that fit in int and
fold at compile time (64 * 1024, 2 * 1024 * 1024, 24 * 60 * 60), so they are
config rather than code: IgnoreConstantIntExpr. The other twelve are runtime
int products widened to an index, and take a cast. unused-return-value is
AllowCastToVoid, since static_cast<void>(p.release()) is how both sites
already said the discard was deliberate.

The reserved-identifier and return-const-ref NOLINTs were both dead. Each sat
above a comment that wrapped onto a second line, so it suppressed the comment
rather than the code, and return-const-ref reports on the return statement
rather than on the declaration the marker sat above. The four deliberate
ClipWaveformPainter roundings now go through one named helper carrying a
single suppression, instead of four.

The five checks #2529 added to the gate's exclusion list are removed: no
check by any of those names exists. The real spellings are
hicpp-signed-bitwise, cert-flp30-c and cert-err34-c, none of them reachable
from a bugprone-* allow-list, so the lines excluded nothing and the sample
counts recorded against them measured something else.

+++ magda/daw/ui/components/chain/custom_ui/FMUI.cpp
@@ -309,8 +309,9 @@ void FMUI::resized() {
         for (int src = 0; src < kNumOps; ++src)
             for (int dst = 0; dst < kNumOps; ++dst) {
                 juce::Rectangle<int> cell(a.getX() + dst * cw, a.getY() + src * chh, cw, chh);
-                controls_[static_cast<size_t>(kMatrixBase) + src * kNumOps + dst].slider->setBounds(
-                    cell.reduced(kCellPad));
+                controls_[static_cast<size_t>(kMatrixBase) + static_cast<size_t>(src) * kNumOps +
+                          dst]
+                    .slider->setBounds(cell.reduced(kCellPad));
             }
     }
